chore: add back in shared switch
